Update external links to have static text
External links are currently problematic as the link text cannot be checked in templates - it's always defined if the link value is defined. This means we can't find out if the title is present (and use that for the text) or empty (else, use 'Learn More') - content.field_link[0]['#title'] is always filled if content.field_link[0]['#url'] is filled (will use url in this variable if it's not otherwise defined).